+++ sdk/src/main/scala/com/akkaserverless/javasdk/impl/valueentity/ResolvedValueEntityFactory.scala
@@ -18,7 +18,7 @@ package com.akkaserverless.javasdk.impl.valueentity
 
 import com.akkaserverless.javasdk.impl.ResolvedEntityFactory
 import com.akkaserverless.javasdk.impl.ResolvedServiceMethod
-import com.akkaserverless.javasdk.lowlevel.ValueEntityFactory
+import com.akkaserverless.javasdk.impl.ValueEntityFactory
 import com.akkaserverless.javasdk.valueentity.ValueEntityContext
 
 class ResolvedValueEntityFactory(delegate: ValueEntityFactory,
