+++ sdk/src/main/scala/com/akkaserverless/javasdk/impl/view/ViewsImpl.scala
@@ -17,13 +17,15 @@
 package com.akkaserverless.javasdk.impl.view
 
 import java.util.Optional
+
 import scala.compat.java8.OptionConverters._
 import scala.util.control.NonFatal
+
 import akka.actor.ActorSystem
 import akka.stream.scaladsl.Source
+import com.akkaserverless.javasdk.impl.ViewFactory
 import com.akkaserverless.javasdk.{Context, Metadata, Service, ServiceCallFactory}
 import com.akkaserverless.javasdk.impl._
-import com.akkaserverless.javasdk.lowlevel.ViewFactory
 import com.akkaserverless.javasdk.view.ViewCreationContext
 import com.akkaserverless.javasdk.view.{UpdateContext, View, ViewContext}
 import com.akkaserverless.protocol.{view => pv}
